Literate programming and CMake.
This commit introduces an approach to use literate programming (using a
narrative to describe the rationale of the code as it is being
developed) along with actual tests. What we accomplish in this commit
are the following:

1) Provide a narrative on what the though process is in the
implementation of the chain type. We want this to be able to go back to
the rationale in the history and evolve the implementation as we're
going along merrily.

2) Implement a way of using the C++0x features in the build
configuration so that we can test appropriately.

The test in here doesn't link yet because there's a lot of stubbed out
function template definitions in the chain implementation. We want to be
able to work through the implementation while amending the rationale and
getting this to a point where the comments can actually be processed to
create a document.