Exposition on the value semantics implementation.
In this commit we basically expound on the case for when we want the
implementation to guarantee exception safety when performing operations
with stateful allocators. This means we get one way of implementing an
exception-robust swap, copy, and assignment.

This is a continuation of the literate exploration of the implementation
of the chain type.